Move gene-related code into genetic module, not components
